+++ apps/remix-app/api/index.js
@@ -1,0 +1,6 @@
+// @ts-check
+const { createRequestHandler } = require('@remix-run/vercel');
+
+module.exports = createRequestHandler({
+  build: require('./build'),
+});