+++ apps/web-app/src/pages/api/hello.ts
@@ -1,7 +1,7 @@
 import { sayHello } from '@your-org/core-lib';
 import type { NextApiRequest, NextApiResponse } from 'next';
 
-export default async function ApiHelloRoute(
+export default async function handleApiHelloRoute(
   req: NextApiRequest,
   res: NextApiResponse
 ) {