+++ packages/ui-lib/.eslintrc.js
@@ -1,16 +1,14 @@
 module.exports = {
   root: true,
-  ignorePatterns: ['dist', 'build'],
   extends: [
     '../../.eslintrc.base.js',
-    // Add specific rules for react
     'plugin:react/recommended',
     'plugin:react-hooks/recommended',
     'plugin:jsx-a11y/recommended',
   ],
   // By loading testing-library as a plugin, we can only enable it
   // on test files via overrides.
-  plugins: ['testing-library'],
+  plugins: ['testing-library', 'storybook'],
   env: {
     browser: true,
     es6: true,
@@ -26,6 +24,11 @@ module.exports = {
       // For performance run jest/recommended on test files, not regular code
       files: ['**/__tests__/**/*.{ts,tsx}'],
       extends: ['plugin:testing-library/react'],
+    },
+    {
+      // For performance run storybook/recommended on test files, not regular code
+      files: ['**/*.stories.{ts,tsx,mdx}'],
+      extends: ['plugin:storybook/recommended'],
     },
   ],
 };