+++ apps/vite-app/CHANGELOG.md
@@ -1,12 +1,31 @@
 # vite-app
 
+## 0.2.0
+
+### Minor Changes
+
+- [#963](https://github.com/belgattitude/nextjs-monorepo-example/pull/963) [`5e25fab2`](https://github.com/belgattitude/nextjs-monorepo-example/commit/5e25fab2f0d620e999f536a1fb8e0ef45d56fd64) Thanks [@belgattitude](https://github.com/belgattitude)! - Upgraded to react 18.0.0-rc.0
+
+### Patch Changes
+
+- Updated dependencies [[`5e25fab2`](https://github.com/belgattitude/nextjs-monorepo-example/commit/5e25fab2f0d620e999f536a1fb8e0ef45d56fd64)]:
+  - @your-org/core-lib@3.9.0
+  - @your-org/ui-lib@3.8.0
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`cfcab664`](https://github.com/belgattitude/nextjs-monorepo-example/commit/cfcab66479a8b28468f67748abb559c4eb2fb10a)]:
+  - @your-org/ui-lib@3.7.0
+
 ## 0.1.0
 
 ### Minor Changes
 
 - [#696](https://github.com/belgattitude/nextjs-monorepo-example/pull/696) [`b04ad38`](https://github.com/belgattitude/nextjs-monorepo-example/commit/b04ad38438d71a24f28d3eab8b34a8737e55232b) Thanks [@belgattitude](https://github.com/belgattitude)! - Add eslint / typecheck / ci for vite-app
 
-- [#661](https://github.com/belgattitude/nextjs-monorepo-example/pull/661) [`0334541`](https://github.com/belgattitude/nextjs-monorepo-example/commit/0334541967b3d13451e6b3eed54d95ea788f5910) Thanks [@ekafyi](https://github.com/ekafyi)! - Vite app example added
+* [#661](https://github.com/belgattitude/nextjs-monorepo-example/pull/661) [`0334541`](https://github.com/belgattitude/nextjs-monorepo-example/commit/0334541967b3d13451e6b3eed54d95ea788f5910) Thanks [@ekafyi](https://github.com/ekafyi)! - Vite app example added
 
 ### Patch Changes